+++ MiniProject/Assets/01.Script/Data/MonsterSkillData.cs
@@ -1,13 +1,13 @@
 ﻿using GlobalDefine;
 public class MonsterSkillData : TableBase
 {
-	public float skillID;
+	public int skillID;
 	public string skillName;
 	public string decal;
 	public eAttackType skillType;
 	public eSkillType target;
 	public int[] optionArr;
-	public override float GetTableID()
+	public override int GetTableID()
 	{
 		return skillID;
 	}
